two photos
The lighting of URP makes 2D games look great and exquisite, and it goes further and further in adjusting UI and sound effects
How about doing a few more levels in the future? If the difficulty increases further, how about adding a bunch of kittens and following Cat Man?
Of course, the resources are temporarily sourced from the Unity store, and the ideas are still my own!